# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MYiTH Core API
  description: |
    채용공고와 국가직무능력표준(NCS)을 결합한 개인 맞춤형 취업 로드맵 서비스.

    인증: Google ID Token으로 로그인 후 발급받은 accessToken을
    Authorization: Bearer {token} 헤더에 실어 보낸다.

    공통 응답: 모든 성공 응답은 data로 감싼다. (GET /api/health 제외)
    공통 오류: { "error": { code, message, fieldErrors, requestId } }
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: 로컬
- url: https://api.myith.example
  description: 운영
security:
- bearerAuth: []
...
paths:
...
  /api/jobs/{jobCode}/diagnosis:
    get:
      tags:
      - Catalog
      summary: 자가진단 문항 조회
      description: |-
        화면 3(자가진단) 문항 목록을 반환한다.
        문항 개수는 직무마다 다르다(6~10개). 배열 길이를 고정하지 않는다.
        levels는 4단계 고정이다. mastery는 참고용이며 프론트는 id만 서버로 보낸다.
        profileVersion은 프론트가 보관했다가 POST /api/roadmaps 요청에 그대로 실어 보낸다.
      operationId: getDiagnosis
      parameters:
      - name: jobCode
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              example:
                data:
                  jobCode: server
                  profileVersion: 1
                  questions:
                  - skillCode: git
                    text: 버전관리로 협업한다
                    axisCode: collaboration
                    axisName: 협업·형상관리
                    sortOrder: 1
                  - skillCode: rest
                    text: REST API 서버를 구현한다
                    axisCode: server-api
                    axisName: 서버·API
                    sortOrder: 6
                  levels:
                  - id: unknown
                    label: 모름
                    mastery: 0
                  - id: heard
                    label: 들어봄
                    mastery: 0.33
                  - id: tried
                    label: 해봄
                    mastery: 0.66
                  - id: independent
                    label: 혼자 가능
                    mastery: 1.0
        "404":
          description: JOB_PROFILE_NOT_READY — 해당 직무의 job_profile 미생성
          content:
            application/json:
              example:
                error:
                  code: JOB_PROFILE_NOT_READY
                  message: 해당 직무의 프로필이 아직 준비되지 않았습니다.
                  requestId: req_01J3ABC
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT